Enable XRD generation alongside CRDs
This commit introduces changes necessary to support XRD to Api docs
compilation.

It achieves this by detecting both CRDs and XRDs, which are simply an
extended CRD format, and if an XRD is detected, loading both formats;
this allows for minimal changes to the code structure whilst still
supporting XRDs.

Additional properties are added to the go template attributes fields to
support the extended mode.

See the template https://github.com/giantswarm/crossplane-gs-apis/blob/main/doc-gen.config
for an example of the fields output.